refactor(agent-core): 统一 Azure resource helper 至 agent-core
extractAzureResourceNameFromApiBase 等 helper 迁入 agent-core/azure-resource；host-internal 与 responses-compat 改为复用单一实现。
